Proxy events to the service, and handle WS
This adds slightly rough implementations for
 - proxying events to the upstream service (Weave Cloud)
 - accepting websocket connections

The latter is to support present fluxd, which will attempt to connect
a websocket to the service given in the arg `--connect`, as well as
send events to it. Since the websocket is intended for RPC _from_ the
service, all the implementation needs to do is pull packets through it
so pings are ponged.