Use a local name in Makefile for the image
To be able to be clearer about which image is which, use a dummy,
local name for the image being built in the Makefile. That can be
referred to directly (e.g., when running in minikube), or retagged to
be pushed somewhere.

The CircleCI script does the latter, firstly naming the image after
the commit SHA1, then retagging it and pushing that to DockerHub.