Add basic proxying Docker image
As a web UI, Weave Cloud will send through the same request as many
times as there are tabs open, but much of the time the request (and
response) are exactly the same.

To coalesce them, we can put a caching proxy between the adapter and
the flux daemon. This commit adds an image build that uses Varnish for
this proxy, with a very minimal configuration.